% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit.R
\name{add_class}
\alias{add_class}
\title{Add a class to the root node of an HTML string}
\usage{
add_class(html, class_name)
}
\arguments{
\item{html}{Character. An HTML string.}

\item{class_name}{Character. Class name to be added to the root node.}
}
\value{
Character. An HTML string with the class name added to the root node.
}
\description{
Add a class to the root node of an HTML string
}
\examples{
library(magrittr)

html <- "<!-- function handles comments--><div><p>R is fun</div></p>"
html \%>\% add_class("second-class")
html \%>\% add_class(".ok-to-start-with-period")
html \%>\% add_class("you-can") \%>\% add_class("chain-function-calls-too")
html \%>\% add_class("or add five at once")

html <- "<div class='using-single-quotes'><p>R is fun</div></p>"
html \%>\% add_class("is-ok-too")

html <- "<div class = 'lots-of-spaces'><p>R is fun</div></p>"
html \%>\% add_class("is-also-ok")
}
