% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sassy.R
\name{add_dependency}
\alias{add_dependency}
\title{Add a dependency}
\usage{
add_dependency(name, version, file = NULL, path = NULL,
  packages = c("rmarkdown", "xaringan", "surprisinglytidy"),
  fullpath = NULL, attach = TRUE)
}
\arguments{
\item{name}{See \code{htmltools::htmlDependency}.}

\item{version}{See \code{htmltools::htmlDependency}.}

\item{file}{Filename of script or stylesheet.}

\item{path}{Path within the package.
Ignored if \code{fullpath} is specified.}

\item{packages}{A vector of package names to search for the file.
Ignored if \code{fullpath} is specified.}

\item{fullpath}{Full path (relative or absolute) of file (all but filename).}

\item{attach}{If TRUE, attaches the dependency.
If FALSE, returns the htmlDependency object to be attached.}
}
\description{
Adds a dependency (external reference) to an R Markdown document.
}
\details{
Inspired by \href{https://www.jaredlander.com/2017/07/fullscreen-background-images-in-ioslides-presentations/}{this guy}.
Must specify \code{file} (for package system files) or \code{fullpath}.
\code{path} is optional, to specify a path for a package system file.
}
\examples{
\dontrun{
add_dependency(
  "testy-css",
  packageVersion("surprisinglytidy"),
  file = "testy.css",
  path = "rmd/h/css"
)
add_dependency(
  "foo",
  "0.0.1",
  fullpath = "./bar.js" # same folder as .Rmd file
)
}
}
